adds support for windows 10 UWP apps
