+++ plugin.xml
@@ -58,10 +58,11 @@
 
         <source-file src="src/android/Printer.java" target-dir="src/de/appplant/cordova/plugin/printer" />
     </platform>
-    <!-- windows8 -->
-    <platform name="windows8">
-        <js-module src="src/windows8/PrinterProxy.js" name="PrinterProxy">
-            <clobbers target="" />
+
+    <!-- windows -->
+    <platform name="windows">
+        <js-module src="src/windows/PrinterProxy.js" name="PrinterProxy">
+            <merges target="" />
         </js-module>
     </platform>
